Move _components
